+++ src/main/java/edu/illinois/library/cantaloupe/cache/JdbcCache.java
@@ -48,7 +48,7 @@ class JdbcCache implements DerivativeCache {
     /**
      * Wraps a {@link Blob} OutputStream, for writing an image to a BLOB.
      * The constructor creates a transaction that is committed on close if the
-     * stream is {@link CompletableOutputStream#isCompletelyWritten()
+     * stream is {@link CompletableOutputStream#isComplete()
      * completely written}.
      */
     private class ImageBlobOutputStream extends CompletableOutputStream {
@@ -71,7 +71,16 @@ class JdbcCache implements DerivativeCache {
 
             connection.setAutoCommit(false);
 
-            blob = connection.createBlob();
+            final Configuration config = Configuration.getInstance();
+            final String sql = String.format(
+                    "INSERT INTO %s (%s, %s, %s) VALUES (?, ?, ?)",
+                    config.getString(Key.JDBCCACHE_DERIVATIVE_IMAGE_TABLE),
+                    DERIVATIVE_IMAGE_TABLE_OPERATIONS_COLUMN,
+                    DERIVATIVE_IMAGE_TABLE_IMAGE_COLUMN,
+                    DERIVATIVE_IMAGE_TABLE_LAST_ACCESSED_COLUMN);
+            LOGGER.trace(sql);
+
+            final Blob blob = connection.createBlob();
             blobOutputStream = blob.setBinaryStream(1);
         }
 
@@ -80,7 +89,7 @@ class JdbcCache implements DerivativeCache {
             LOGGER.debug("Closing stream for {}", ops);
             PreparedStatement statement = null;
             try {
-                if (isCompletelyWritten()) {
+                if (isComplete()) {
                     blobOutputStream.close();
                     final Configuration config = Configuration.getInstance();
                     final String sql = String.format(
@@ -253,7 +262,7 @@ class JdbcCache implements DerivativeCache {
             statement.setTimestamp(1, now());
             statement.setString(2, opList.toString());
 
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             statement.executeUpdate();
         }
     }
@@ -288,7 +297,7 @@ class JdbcCache implements DerivativeCache {
             statement.setTimestamp(1, now());
             statement.setString(2, identifier.toString());
 
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             statement.executeUpdate();
         }
     }
@@ -332,7 +341,7 @@ class JdbcCache implements DerivativeCache {
             statement.setString(1, identifier.toString());
             statement.setTimestamp(2, earliestValidDate());
 
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
                     accessInfoAsync(identifier);
@@ -368,7 +377,7 @@ class JdbcCache implements DerivativeCache {
             statement.setString(1, opList.toString());
             statement.setTimestamp(2, earliestValidDate());
 
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
                     LOGGER.debug("Hit for image: {}", opList);
@@ -432,6 +441,23 @@ class JdbcCache implements DerivativeCache {
     }
 
     @Override
+    public void purgeInfos() throws IOException {
+        try (Connection connection = getConnection()) {
+            connection.setAutoCommit(false);
+            int numDeleted;
+            final String sql = "DELETE FROM " + getInfoTableName();
+            try (PreparedStatement statement = connection.prepareStatement(sql)) {
+                LOGGER.trace(sql);
+                numDeleted = statement.executeUpdate();
+            }
+            connection.commit();
+            LOGGER.debug("purgeInfos(): purged {} info(s)", numDeleted);
+        } catch (SQLException e) {
+            throw new IOException(e.getMessage(), e);
+        }
+    }
+
+    @Override
     public void purgeInvalid() throws IOException {
         try (Connection connection = getConnection()) {
             connection.setAutoCommit(false);
@@ -457,7 +483,7 @@ class JdbcCache implements DerivativeCache {
                 DERIVATIVE_IMAGE_TABLE_LAST_ACCESSED_COLUMN);
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
             statement.setTimestamp(1, earliestValidDate());
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -472,7 +498,7 @@ class JdbcCache implements DerivativeCache {
                 getInfoTableName(), INFO_TABLE_LAST_ACCESSED_COLUMN);
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
             statement.setTimestamp(1, earliestValidDate());
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -489,7 +515,7 @@ class JdbcCache implements DerivativeCache {
                 DERIVATIVE_IMAGE_TABLE_OPERATIONS_COLUMN);
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
             statement.setString(1, ops.toString());
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -516,7 +542,7 @@ class JdbcCache implements DerivativeCache {
     private int purgeDerivativeImages(Connection conn) throws SQLException {
         final String sql = "DELETE FROM " + getDerivativeImageTableName();
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -536,7 +562,7 @@ class JdbcCache implements DerivativeCache {
                 " LIKE ?";
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
             statement.setString(1, identifier.toString() + "%");
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -570,7 +596,7 @@ class JdbcCache implements DerivativeCache {
                 getInfoTableName(), INFO_TABLE_IDENTIFIER_COLUMN);
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
             statement.setString(1, identifier.toString());
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -592,7 +618,7 @@ class JdbcCache implements DerivativeCache {
     private int purgeInfos(Connection conn) throws SQLException {
         final String sql = "DELETE FROM " + getInfoTableName();
         try (PreparedStatement statement = conn.prepareStatement(sql)) {
-            LOGGER.debug(sql);
+            LOGGER.trace(sql);
             return statement.executeUpdate();
         }
     }
@@ -600,7 +626,7 @@ class JdbcCache implements DerivativeCache {
     @Override
     public void put(Identifier identifier, Info info) throws IOException {
         if (!info.isPersistable()) {
-            LOGGER.debug("put(): info for {} is incomplete; ignoring",
+            LOGGER.trace("put(): info for {} is incomplete; ignoring",
                     identifier);
             return;
         }