+++ src/main/java/edu/illinois/library/cantaloupe/processor/codec/jpeg/Util.java
@@ -1,8 +1,7 @@
 package edu.illinois.library.cantaloupe.processor.codec.jpeg;
 
-import edu.illinois.library.cantaloupe.image.Metadata;
+import edu.illinois.library.cantaloupe.image.xmp.Utils;
 import edu.illinois.library.cantaloupe.util.ArrayUtils;
-import edu.illinois.library.cantaloupe.util.StringUtils;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.RDFNode;
@@ -29,7 +28,7 @@ final class Util {
         try {
             final ByteArrayOutputStream os = new ByteArrayOutputStream();
             final byte[] headerBytes = Constants.STANDARD_XMP_SEGMENT_HEADER;
-            final byte[] xmpBytes = Metadata.encapsulateXMP(xmp).
+            final byte[] xmpBytes = Utils.encapsulateXMP(xmp).
                     getBytes(StandardCharsets.UTF_8);
             // write segment marker
             os.write(Marker.APP1.marker());
@@ -61,12 +60,12 @@ final class Util {
         final int numChunks = xmpChunks.size();
         if (numChunks > 0) {
             standardXMP = new String(xmpChunks.get(0), StandardCharsets.UTF_8);
-            standardXMP = StringUtils.trimXMP(standardXMP);
+            standardXMP = Utils.trimXMP(standardXMP);
             if (numChunks > 1) {
                 String extendedXMP = new String(
                         mergeChunks(xmpChunks.subList(1, numChunks)),
                         StandardCharsets.UTF_8);
-                extendedXMP = StringUtils.trimXMP(extendedXMP);
+                extendedXMP = Utils.trimXMP(extendedXMP);
                 return mergeXMPModels(standardXMP, extendedXMP);
             }
         }