+++ src/test/java/edu/illinois/library/cantaloupe/image/MetadataTest.java
@@ -6,6 +6,7 @@ import edu.illinois.library.cantaloupe.image.exif.Rational;
 import edu.illinois.library.cantaloupe.image.exif.Tag;
 import edu.illinois.library.cantaloupe.image.exif.TagSet;
 import edu.illinois.library.cantaloupe.image.iptc.DataSet;
+import edu.illinois.library.cantaloupe.image.xmp.Utils;
 import edu.illinois.library.cantaloupe.processor.codec.ImageReader;
 import edu.illinois.library.cantaloupe.processor.codec.ImageReaderFactory;
 import edu.illinois.library.cantaloupe.test.BaseTest;
@@ -17,6 +18,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -34,17 +36,11 @@ public class MetadataTest extends BaseTest {
         instance = new Metadata();
     }
 
-    @Test
-    void testEncapsulateXMP() {
-        final String xmp = "<rdf:RDF></rdf:RDF>";
-        String actual = Metadata.encapsulateXMP(xmp);
-        assertTrue(actual.startsWith("<?xpacket"));
-        assertTrue(actual.endsWith("<?xpacket end=\"r\"?>"));
-    }
+    /* equals() */
 
     @Test
     void testEqualsWithEqualInstances() {
-        Directory exif = new Directory(TagSet.EXIF);
+        Directory exif     = new Directory(TagSet.EXIF);
         List<DataSet> iptc = List.of(new DataSet(
                 edu.illinois.library.cantaloupe.image.iptc.Tag.CITY,
                 "Urbana".getBytes()));
@@ -119,9 +115,11 @@ public class MetadataTest extends BaseTest {
         assertNotEquals(m1, m2);
     }
 
+    /* getEXIF() */
+
     @Test
     void testGetEXIFWithPresentEXIFData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg-exif.jpg");
+        Path fixture       = TestUtil.getImage("jpg-exif.jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -134,7 +132,7 @@ public class MetadataTest extends BaseTest {
 
     @Test
     void testGetEXIFWithNoEXIFData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg");
+        Path fixture       = TestUtil.getImage("jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -145,9 +143,11 @@ public class MetadataTest extends BaseTest {
         }
     }
 
+    /* getIPTC() */
+
     @Test
     void testGetIPTCWithPresentIPTCData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg-iptc.jpg");
+        Path fixture       = TestUtil.getImage("jpg-iptc.jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -160,7 +160,7 @@ public class MetadataTest extends BaseTest {
 
     @Test
     void testGetIPTCWithNoIPTCData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg");
+        Path fixture       = TestUtil.getImage("jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -170,6 +170,8 @@ public class MetadataTest extends BaseTest {
             reader.dispose();
         }
     }
+
+    /* getNativeMetadata() */
 
     @Test
     void testGetNativeMetadataWithPresentData() throws Exception {
@@ -249,9 +251,40 @@ public class MetadataTest extends BaseTest {
         assertEquals(Orientation.ROTATE_0, instance.getOrientation());
     }
 
+    /* getXMPElements() */
+
+    @Test
+    void testGetXMPElementsWithPresentXMPData() throws Exception {
+        Path fixture       = TestUtil.getImage("jpg-xmp.jpg");
+        ImageReader reader = new ImageReaderFactory()
+                .newImageReader(Format.get("jpg"), fixture);
+        try {
+            Metadata metadata        = reader.getMetadata(0);
+            Map<String,Object> model = metadata.getXMPElements();
+            assertEquals(6, model.size());
+        } finally {
+            reader.dispose();
+        }
+    }
+
+    @Test
+    void testGetXMPElementsWithNoXMPData() throws Exception {
+        Path fixture       = TestUtil.getImage("jpg");
+        ImageReader reader = new ImageReaderFactory()
+                .newImageReader(Format.get("jpg"), fixture);
+        try {
+            Metadata metadata = reader.getMetadata(0);
+            assertTrue(metadata.getXMPElements().isEmpty());
+        } finally {
+            reader.dispose();
+        }
+    }
+
+    /* getXMPModel() */
+
     @Test
     void testGetXMPModelWithPresentXMPData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg-xmp.jpg");
+        Path fixture       = TestUtil.getImage("jpg-xmp.jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -265,7 +298,7 @@ public class MetadataTest extends BaseTest {
 
     @Test
     void testGetXMPModelWithNoXMPData() throws Exception {
-        Path fixture = TestUtil.getImage("jpg");
+        Path fixture       = TestUtil.getImage("jpg");
         ImageReader reader = new ImageReaderFactory()
                 .newImageReader(Format.get("jpg"), fixture);
         try {
@@ -276,10 +309,12 @@ public class MetadataTest extends BaseTest {
         }
     }
 
+    /* hashCode() */
+
     @Test
     void testHashCodeWithEqualInstances() {
         Directory exif = new Directory(TagSet.EXIF);
-        String xmp = "<rdf:RDF>cats</rdf:RDF>";
+        String xmp     = "<rdf:RDF>cats</rdf:RDF>";
 
         Metadata m1 = new Metadata();
         m1.setEXIF(exif);
@@ -348,17 +383,23 @@ public class MetadataTest extends BaseTest {
         assertNotEquals(m1.hashCode(), m2.hashCode());
     }
 
+    /* setEXIF() */
+
     @Test
     void testSetEXIFWithNullArgument() {
         instance.setEXIF(null);
         assertFalse(instance.getEXIF().isPresent());
     }
 
+    /* setIPTC() */
+
     @Test
     void testSetIPTCWithNullArgument() {
         instance.setIPTC(null);
         assertFalse(instance.getIPTC().isPresent());
     }
+
+    /* setXMP() */
 
     @Test
     void testSetXMPWithNullByteArrayArgument() {
@@ -379,6 +420,8 @@ public class MetadataTest extends BaseTest {
         assertTrue(xmp.startsWith("<rdf:RDF"));
         assertTrue(xmp.endsWith("</rdf:RDF>"));
     }
+
+    /* toMap() */
 
     @Test
     void testToMap() {
@@ -404,6 +447,7 @@ public class MetadataTest extends BaseTest {
                 edu.illinois.library.cantaloupe.image.iptc.Tag.CITY,
                 "Urbana".getBytes()).toMap()));
         expectedMap.put("xmp_string", "<rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\"></rdf:RDF>");
+        expectedMap.put("xmp_elements", Collections.emptyMap());
         expectedMap.put("native", Map.of("key1", "value1", "key2", "value2"));
 
         // assemble the Metadata