+++ src/test/java/edu/illinois/library/cantaloupe/source/HTTPRequestInfoTest.java
@@ -4,12 +4,11 @@ import edu.illinois.library.cantaloupe.test.BaseTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class HTTPRequestInfoTest extends BaseTest {
+class HTTPRequestInfoTest extends BaseTest {
 
     private HTTPRequestInfo instance;
 
@@ -17,22 +16,39 @@ public class HTTPRequestInfoTest extends BaseTest {
     public void setUp() throws Exception {
         super.setUp();
 
-        Map<String,String> headers = new HashMap<>();
-        headers.put("X-Animal", "cats");
-
-        instance = new HTTPRequestInfo("http://example.org/cats",
-                "user", "secret", headers);
+        instance = new HTTPRequestInfo();
+        instance.setURI("http://example.org/cats");
+        instance.setUsername("user");
+        instance.setSecret("secret");
+        instance.setHeaders(Map.of("X-Animal", "cats"));
+        instance.setSendingHeadRequest(true);
     }
 
     @Test
-    void testGetBasicAuthTokenWithoutUserAndSecret() {
-        instance = new HTTPRequestInfo("http://example.org/cats");
+    void getBasicAuthTokenWithoutUserAndSecret() {
+        instance = new HTTPRequestInfo();
+        instance.setURI("http://example.org/cats");
         assertNull(instance.getBasicAuthToken());
     }
 
     @Test
-    void testGetBasicAuthTokenWithUserAndSecret() {
+    void getBasicAuthTokenWithUserAndSecret() {
         assertEquals("dXNlcjpzZWNyZXQ=", instance.getBasicAuthToken());
+    }
+
+    @Test
+    void setHeaders() {
+        assertEquals(1, instance.getHeaders().size());
+
+        instance.setHeaders(Map.of("X-Cats", "yes"));
+        assertEquals(2, instance.getHeaders().size());
+        assertEquals("yes", instance.getHeaders().getFirstValue("X-Cats"));
+    }
+
+    @Test
+    void setHeadersWithNullArgument() {
+        instance.setHeaders(null);
+        assertEquals(0, instance.getHeaders().size());
     }
 
 }