+++ src/test/resources/delegates.rb
@@ -182,14 +182,16 @@ class CustomDelegate
             'uri' => 'http://example.org/bla/' + CGI.escape(identifier),
             'headers' => {
                 'X-Custom' => 'yes'
-            }
+            },
+            'send_head_request' => true
         }
       when 'https-jpg-rgb-64x56x8-baseline.jpg'
         return {
             'uri' => 'https://example.org/bla/' + CGI.escape(identifier),
             'headers' => {
                 'X-Custom' => 'yes'
-            }
+            },
+            'send_head_request' => true
         }
       when 'http-jpg-rgb-64x56x8-plane.jpg'
         return {
@@ -198,7 +200,8 @@ class CustomDelegate
             'secret' => 'secret',
             'headers' => {
                 'X-Custom' => 'yes'
-            }
+            },
+            'send_head_request' => true
         }
       when 'https-jpg-rgb-64x56x8-plane.jpg'
         return {
@@ -207,7 +210,8 @@ class CustomDelegate
             'secret' => 'secret',
             'headers' => {
                 'X-Custom' => 'yes'
-            }
+            },
+            'send_head_request' => true
         }
     end
     nil
@@ -215,6 +219,10 @@ class CustomDelegate
 
   def jdbcsource_database_identifier(options = {})
     context['identifier']
+  end
+
+  def jdbcsource_last_modified(options = {})
+    'SELECT last_modified FROM items WHERE filename = ?'
   end
 
   def jdbcsource_media_type(options = {})